requirements/requirements.txt
# Python requirements
psycopg2-binary
python-dotenv
alembic
sqlalchemy
